+++ events/messageCreate.js
@@ -1,4 +1,3 @@
-const { MessageEmbed } = require('discord.js')
 const { cooldown } = require('../functions/commands')
 const { Status, Error } = require('../modules/constructor/messageBuilder')
 const getLocales = require('../i18n/getLocales')
@@ -6,6 +5,7 @@ const autoresponder = require('../modules/autoresponder')
 const guildFetchData = require('../functions/guildFetchData')
 const { rankUp } = require('../modules/levels')
 const humanizeduration = require('humanize-duration')
+const customcommands = require('../modules/customcommands')
 
 module.exports = async (client, message) => {
   if (
@@ -47,34 +47,7 @@ module.exports = async (client, message) => {
       } else {
         if (cooldown.check(message.member, message.guild, commandToExecute)) {
           cooldown.add(message.member, message.guild, commandToExecute)
-          const mCeEC = client.Sentry.startTransaction({
-            op: 'messageCreate/executeExternalCommand',
-            name: 'Execute External Command'
-          })
-          client.pool.query('SELECT * FROM `guildCustomCommands` WHERE `guild` = ?', [message.guild.id], (err, result) => {
-            if (err) {
-              client.Sentry.captureException(err)
-              client.log.error(err)
-            }
-            if (Object.prototype.hasOwnProperty.call(result, 0)) {
-              client.pool.query('SELECT * FROM `guildCustomCommands` WHERE `guild` = ? AND `customCommand` = ?', [message.guild.id, commandToExecute], (err, result) => {
-                if (err) {
-                  client.Sentry.captureException(err)
-                  client.log.error(err)
-                }
-                if (Object.prototype.hasOwnProperty.call(result, 0)) {
-                  const messageSent = new MessageEmbed()
-                    .setFooter('Powered by Pingu', 'https://cdn.discordapp.com/attachments/907917245567598592/907917308620587059/Instagram_Profiles1.png')
-                    .setDescription(result[0].messageReturned)
-                    .setColor('BLURPLE')
-                  message.channel.send({ embeds: [messageSent] }).catch((err) => {
-                    client.log.error(err)
-                    client.Sentry.captureException(err)
-                  }).finally(mCeEC.finish())
-                }
-              })
-            }
-          })
+          customcommands(client, message, commandToExecute)
         } else {
           message.reply({ embeds: [Error(getLocales(message.database.guildLanguage || 'en', 'COOLDOWN', { COOLDOWN: humanizeduration(cooldown, { round: true, language: message.database.guildLanguage || 'en', fallbacks: ['en'] }) }))] })
           return
